+++ deployments/bin/delete_instance.sh
@@ -35,11 +35,9 @@ terraform apply ${TF_INPUT_FALSE} ${TF_AUTO_APPROVE} -target=module.pim.local_fi
 echo "2 - removing deployment and terraform resources"
 export KUBECONFIG=.kubeconfig
 
-# WARNING ! DON'T DELETE release helm before get list of PV
-# Empty list is not an error
+# WARNING ! DON'T DELETE release helm before get list of PD
 # grep -v mysql because the mysql disk is manage by terraform process
 LIST_PD_NAME=$((kubectl get pv -o json | jq -r --arg PFID "$PFID" '[.items[] | select(.spec.claimRef.namespace == $PFID) | .spec.gcePersistentDisk.pdName] | unique | .[]' | grep -v mysql) || echo "")
-LIST_PV_NAME=$((kubectl get pv -o json | jq -r --arg PFID "$PFID" '[.items[] | select(.spec.claimRef.namespace == $PFID) | .metadata.name] | unique | .[]' | grep -v mysql) || echo "")
 
 (helm3 list -n "${PFID}" && helm3 uninstall ${PFID} -n ${PFID}) || true
 ((kubectl get ns ${PFID} | grep "$PFID") && kubectl delete ns ${PFID}) || true
@@ -57,14 +55,21 @@ sleep 30
 gsutil rm -r gs://akecld-terraform${TF_BUCKET}/saas/${GOOGLE_PROJECT_ID}/${GOOGLE_CLUSTER_ZONE}/${PFID}
 
 echo "4 - Delete PD and PV"
+# Check disk still exist
 if [[ -n "${LIST_PD_NAME}" ]]; then
   for PD_NAME in ${LIST_PD_NAME}; do
+    IS_DISK_DETACHED=$(gcloud --project=${GOOGLE_PROJECT_ID} compute disks list --filter="(name=(${PD_NAME}) AND zone:${GOOGLE_CLUSTER_ZONE} AND NOT users:*)" --format="value(name)" )
+    if [[ -z "$IS_DISK_DETACHED" ]]; then
+      break;
+    fi
     for i in {1..6}; do
   		gcloud --quiet compute disks delete ${PD_NAME} --project=${GOOGLE_PROJECT_ID} --zone=${GOOGLE_CLUSTER_ZONE} && break || sleep 10
   	done
   done
 fi
 
+# Empty list is not an error
+LIST_PV_NAME=$((kubectl get pv -o json | jq -r --arg PFID "$PFID" '[.items[] | select(.spec.claimRef.namespace == $PFID) | .metadata.name] | unique | .[]') || echo "")
 if [[ -n "${LIST_PV_NAME}" ]]; then
   for PV_NAME in ${LIST_PV_NAME}; do
     kubectl delete pv ${PV_NAME}